refactor(PrestaShop::Spider): add parent class
