+++ spiders/diverti_spider.rb
@@ -1,8 +1,7 @@
 # frozen_string_literal: true
 
 # Diverti store spider
-# engine: Prestashop
-class DivertiSpider < ApplicationSpider
+class DivertiSpider < EcommerceEngines::PrestaShop::Spider
   @name = "diverti_spider"
   @store = {
     name: "Diverti",
@@ -10,28 +9,6 @@ class DivertiSpider < ApplicationSpider
   }
   @start_urls = ["https://www.diverti.cl/juegos-de-mesa-16?en-stock=1"]
   @config = {}
-
-  def parse(response, url:, data: {})
-    items = parse_index(response, url:)
-    items.each { |item| send_item item }
-
-    paginate(response, url)
-  end
-
-  def parse_index(response, url:, data: {})
-    listings = response.css("div#js-product-list article")
-    listings.map { |listing| parse_product_node(listing, url:) }
-  end
-
-  def parse_product_node(node, url:)
-    {
-      url: get_url(node),
-      title: get_title(node),
-      price: get_price(node),
-      stock: purchasable?(node),
-      image_url: get_image_url(node)
-    }
-  end
 
   def next_page_url(response, url)
     # This store doesn't disable the next page link on the last pagination result
@@ -43,35 +20,7 @@ class DivertiSpider < ApplicationSpider
 
   private
 
-  def paginate(response, url)
-    next_page_url = next_page_url(response, url)
-    request_to(:parse, url: next_page_url) if next_page_url
-  end
-
-  def get_url(node)
-    node.at_css(".product-title a")[:href]
-  end
-
-  def get_title(node)
-    node.at_css(".product-title").text.strip
-  end
-
-  def get_price(node)
-    price_node = node.at_css("span.price")
-    scan_int(price_node.text) if price_node
-  end
-
   def in_stock?(node)
     node.at_css("div.ago").text.strip.empty?
-  end
-
-  def purchasable?(node)
-    in_stock?(node)
-  end
-
-  def get_image_url(node)
-    node.at_css("img")["data-full-size-image-url"]
-  rescue NoMethodError
-    nil
   end
 end
